requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
scikit-learn>=1.4.2
spacy>=3.7.4
pymupdf>=1.24.0
python-docx>=1.1.2
pydantic>=2.7.1
sqlalchemy>=2.0.30
gunicorn>=22.0.0
numpy>=1.24.0,<2.0.0
groq==0.9.0
